Error handling improvements.
