+++ node/rpc/src/rpc_http_handler.rs
@@ -1,3 +1,4 @@
+use jsonrpc_v2::RequestObject as JsonRpcRequestObject;
 use tide::http::{format_err, Error as HttpError, Method};
 use tide::{Request as HttpRequest, Response as HttpResponse, Result};
 
@@ -6,8 +7,7 @@ use blockstore::BlockStore;
 use wallet::KeyStore;
 
 use crate::data_types::JsonRpcServerState;
-use crate::rpc_util::is_streaming_method;
-use rpc_types::JsonRpcRequestObject;
+use crate::rpc_util::{is_streaming_method, make_rpc_call};
 
 pub async fn rpc_http_handler<DB, KS, B>(
     mut http_request: HttpRequest<JsonRpcServerState>,
@@ -34,16 +34,15 @@ where
 
     let rpc_call: JsonRpcRequestObject = http_request.body_json().await?;
 
-    if is_streaming_method(rpc_call.method.to_string()) {
+    if is_streaming_method(rpc_call.method_ref()) {
         return Err(HttpError::from_str(
             500,
-            "This endpoint should not handle streaming methods",
+            "This endpoint cannot handle streaming methods",
         ));
     }
 
-    let http_request_bytes = http_request.body_bytes().await.unwrap();
     let rpc_server = http_request.state();
-    let rpc_response = rpc_server.handle(http_request_bytes.as_ref()).await;
+    let rpc_response = make_rpc_call::<_, DB, KS, B>(rpc_server.clone(), rpc_call).await?;
     let http_response = HttpResponse::builder(200)
         .body(serde_json::to_string(&rpc_response).unwrap())
         .content_type("application/json-rpc;charset=utf-8")
